feat(#520): move classes into appropriate packages
