+++ src/main/java/org/eolang/jeo/representation/xmir/HexString.java
@@ -25,7 +25,7 @@ package org.eolang.jeo.representation.xmir;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
-import org.eolang.jeo.representation.DataType;
+import org.eolang.jeo.representation.bytecode.DataType;
 
 /**
  * Hex string.
