+++ src/test/java/org/eolang/jeo/representation/directives/DirectivesAnnotationsTest.java
@@ -25,7 +25,6 @@ package org.eolang.jeo.representation.directives;
 
 import com.jcabi.matchers.XhtmlMatchers;
 import org.eolang.jeo.matchers.SameXml;
-import org.eolang.jeo.representation.HexData;
 import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.Test;
 import org.xembly.ImpossibleModificationException;
