+++ src/test/java/org/eolang/jeo/representation/directives/DirectivesEnumAnnotationValueTest.java
@@ -24,7 +24,7 @@
 package org.eolang.jeo.representation.directives;
 
 import com.jcabi.matchers.XhtmlMatchers;
-import org.eolang.jeo.representation.DataType;
+import org.eolang.jeo.representation.bytecode.DataType;
 import org.eolang.jeo.representation.bytecode.BytecodeEnumAnnotationValue;
 import org.eolang.jeo.representation.xmir.XmlAnnotationProperty;
 import org.eolang.jeo.representation.xmir.XmlNode;
