+++ src/test/java/org/eolang/jeo/representation/directives/DirectivesValueTest.java
@@ -24,7 +24,7 @@
 package org.eolang.jeo.representation.directives;
 
 import org.eolang.jeo.matchers.SameXml;
-import org.eolang.jeo.representation.DataType;
+import org.eolang.jeo.representation.bytecode.DataType;
 import org.eolang.jeo.representation.xmir.AllLabels;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
